+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[unreleased]: https://github.com/johnarp/github-profile-viewer/compare/v0.1.0...HEAD
+[0.2.0]: https://github.com/johnarp/github-profile-viewer/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/johnarp/github-profile-viewer/releases/tag/v0.1.0
 
-## [Unreleased]
+## [0.2.0] - 2026-03-17
+
+### Added
+
+- Enter key to search
+- Recent repositories, joined date, view on github
 
 ### Fixed